requirements.txt
PyPDF2
pdfplumber
Pillow
pdf2image
pytesseract
pdfminer.six
